Drop rounding from HPA queries
Without this change, the rounding can hide data when the values are
below 0.5.  For example, a ScaleObject which scales when the CPU
AverageValue is 300m CPU, or 0.3, the value is rounded to 0, which
appears to have no data on the panels.

Here we remove the rounding from the two panels which this shows up in
most so that data, even in small values, is visible to the user.